+++ CHANGES.md
@@ -11,6 +11,9 @@
 -   Add rudimentary burst event workflow to orchestrator: it just generates
     VOEvents and circulars.
 
+-   create a label in gracedb whenever `em_bright` or `bayestar`
+    completes
+
 ## 0.0.14 (2017-06-28)
 
 -   Fix typo that was causing a task to fail.